Migrated to SpringBoot 2.0.0

+++ backend/src/main/java/org/simonscode/photoshare/repositories/OffsetLimitPagable.java
@@ -32,7 +32,7 @@ public class OffsetLimitPagable implements Pageable {
     }
 
     @Override
-    public int getOffset() {
+    public long getOffset() {
         return offset * page;
     }
